Final/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
pymilvus>=2.4.0
sentence-transformers>=3.0.0
mcp>=1.0.0
markdown>=3.6
python-dotenv>=1.0.0
